pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai_digest"
version = "0.1.0"
description = "Privacy-first AI intelligence digest system: ingest from multiple sources, evaluate with local LLMs, deliver via email/Telegram"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [{name = "AI Digest", email = "support@example.com"}]
keywords = ["ai", "digest", "llm", "local", "privacy", "prometheus", "grafana"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Web Environment",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
    "Topic :: Office/Business",
]
dependencies = [
    "flask>=2.2",
    "pydantic>=2.0",
    "scikit-learn>=1.2",
    "pytest>=7.0",
    "requests>=2.28",
    "aiohttp>=3.9",
    "python-dotenv>=1.0",
    "tenacity>=8.2",
    "praw>=7.8.1",
    "alembic>=1.10",
    "prometheus-client>=0.16",
    "python-telegram-bot>=20.0",
    "aiosmtplib>=2.3",
]

[project.optional-dependencies]
dev = ["pytest>=7.0"]
faiss = ["faiss-cpu>=1.7.0"]

[project.urls]
Homepage = "https://github.com/example/ai_digest"
Issues = "https://github.com/example/ai_digest/issues"
Documentation = "https://github.com/example/ai_digest/blob/main/README.md"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q"
testpaths = ["tests"]
